combine process modules
