+++ src/rapidfuzz/CMakeLists.txt
@@ -27,19 +27,11 @@ target_link_libraries(fuzz_cpp PRIVATE rapidfuzz::rapidfuzz)
 python_extension_module(fuzz_cpp)
 install(TARGETS fuzz_cpp LIBRARY DESTINATION src/rapidfuzz)
 
-create_cython_target(process_cpp)
-add_library(process_cpp MODULE ${process_cpp})
-target_compile_features(process_cpp PUBLIC cxx_std_17)
-target_include_directories(process_cpp PRIVATE ${RF_CAPI_PATH} ${RF_BASE_DIR}/rapidfuzz)
-target_link_libraries(process_cpp PRIVATE rapidfuzz::rapidfuzz)
-python_extension_module(process_cpp)
-install(TARGETS process_cpp LIBRARY DESTINATION src/rapidfuzz)
-
-create_cython_target(process_cdist_cpp_impl)
-add_library(process_cdist_cpp_impl MODULE ${process_cdist_cpp_impl})
-target_compile_features(process_cdist_cpp_impl PUBLIC cxx_std_17)
-target_include_directories(process_cdist_cpp_impl PRIVATE ${RF_CAPI_PATH} ${RF_BASE_DIR}/rapidfuzz)
-target_link_libraries(process_cdist_cpp_impl PRIVATE Taskflow::Taskflow rapidfuzz::rapidfuzz)
+create_cython_target(process_cpp_impl)
+add_library(process_cpp_impl MODULE ${process_cpp_impl})
+target_compile_features(process_cpp_impl PUBLIC cxx_std_17)
+target_include_directories(process_cpp_impl PRIVATE ${RF_CAPI_PATH} ${RF_BASE_DIR}/rapidfuzz)
+target_link_libraries(process_cpp_impl PRIVATE Taskflow::Taskflow rapidfuzz::rapidfuzz)
 
 # check if we need to link with libatomic (not needed on Windows)
 if (NOT Windows)
@@ -74,7 +66,7 @@ if (NOT Windows)
 		else()
 			message(STATUS "Linking with libatomic for atomics support")
 			unset(CMAKE_REQUIRED_LIBRARIES)
-			target_link_libraries(process_cdist_cpp_impl PUBLIC atomic)
+			target_link_libraries(process_cpp_impl PUBLIC atomic)
 		endif()
 	endif()
 	if(APPLE)
@@ -82,5 +74,5 @@ if (NOT Windows)
 	endif()
 endif()
 
-python_extension_module(process_cdist_cpp_impl)
-install(TARGETS process_cdist_cpp_impl LIBRARY DESTINATION src/rapidfuzz)
+python_extension_module(process_cpp_impl)
+install(TARGETS process_cpp_impl LIBRARY DESTINATION src/rapidfuzz)
