+++ src/rapidfuzz/process.py
@@ -7,4 +7,4 @@ _mod = "rapidfuzz.process"
 extract = _fallback_import(_mod, "extract")
 extractOne = _fallback_import(_mod, "extractOne")
 extract_iter = _fallback_import(_mod, "extract_iter")
-cdist = _fallback_import("rapidfuzz.process_cdist", "cdist")
+cdist = _fallback_import(_mod, "cdist")
